+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- hardis:org:diagnose:audittrail: Fix lastndays not taken in account in some contexts
+
 ## [4.32.2] 2024-05-01
 
 - Fix GitHub Actions check deploy workflow